Tests: extract the Account class into its own file
